Add reviewer field and insert briefing quote from original assessment
Reviewer: Needs to be someone OTHER than the assessor. The university insists
on this being in there, makes sense we don't get rid of it. (Means there's
one assessor, and one sanity checker, both of whom are accountable.

Briefing: I believe this statement needs to be here, means we are responsible
to bring risks to attention of the competitors at the start of the day.